Update main.c
The accelerometer actually outputs 3 different values to account for 3 different axes. One of them is the z (up and down) direction, which is always a value around -9.8, which is the gravitational acceleration constant; this is not needed in our calculations. We took the other two accelerations, and used pythagorean’s theorem to find the magnitude of the acceleration when considering both accelerations in the x and y directions.