Update main.c
After reading through the datasheet, it is in fact sent through binary. Thus, we are required to make a change and do a number of binary operations such as bitshift and binary or to actually receive the raw data. In addition, after receiving this raw data, we must decide the values by 16384 in accordance with the accuracy that we have set; we also found the value for this through the datasheet.